♻ More refactor
♻ Moved response cards to their own component
