+++ src/components/Card.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const Card = () => {
+  return <div className="card"></div>;
+};
